pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video2svg"
version = "0.1.0"
description = "Convert videos to animated SVG files"
readme = "README.md"
authors = [
    {name = "Veridock", email = "info@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "opencv-python",
    "numpy",
    "svgwrite",
    "PyYAML",
]
requires-python = ">=3.7"

[project.urls]
"Homepage" = "https://github.com/veridock/video2svg"
"Bug Tracker" = "https://github.com/veridock/video2svg/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["video2svg*"]

[project.scripts]
video2svg = "video2svg.main:main"
